solverstack-crud/requirements.txt
click
Flask
Flask-JWT-Extended
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
python-dotenv
